WIP restructure Auth component
