+++ waspc/data/Generator/templates/react-app/src/auth/forms/core/social/SocialButton.tsx
@@ -1,4 +1,4 @@
-import { styled } from '../../stitches.config'
+import { styled } from '../../../../stitches.config'
 
 export const SocialButton = styled('a', {
     display: 'flex',
