+++ waspc/data/Generator/templates/react-app/src/auth/forms/core/social/SocialIcons.tsx
