+++ waspc/data/Generator/templates/react-app/src/auth/forms/core/usernameAndPassword/useUsernameAndPassword.ts
@@ -1,0 +1,33 @@
+import { useState } from 'react'
+import signup from '../../../signup.js'
+import login from '../../../login.js'
+
+export function useUsernameAndPassword({
+    onError,
+    onSuccess,
+    isLogin,
+}: {
+    onError: (error: Error) => void;
+    onSuccess: () => void;
+    isLogin: boolean;
+}) {
+    const [usernameFieldVal, setUsernameFieldVal] = useState('')
+    const [passwordFieldVal, setPasswordFieldVal] = useState('')
+
+    async function handleSubmit () {
+        try {
+            if (!isLogin) {
+                await signup({ username: usernameFieldVal, password: passwordFieldVal })
+            }
+            await login (usernameFieldVal, passwordFieldVal)
+
+            setUsernameFieldVal('')
+            setPasswordFieldVal('')
+            onSuccess()
+        } catch (err: unknown) {
+            onError(err as Error)
+        }
+    }
+
+    return { handleSubmit, usernameFieldVal, passwordFieldVal, setUsernameFieldVal, setPasswordFieldVal }
+}
