+++ waspc/data/Generator/templates/react-app/src/auth/forms/types.ts
@@ -3,6 +3,7 @@ import { createTheme } from '@stitches/react'
 export enum State {
     Login = "login",
     Signup = "signup",
+    // TODO: adjust if email auth is not used
     ForgotPassword = "forgot-password",
     ResetPassword = "reset-password",
     VerifyEmail = "verify-email",
@@ -13,3 +14,8 @@ export type CustomizationOptions = {
     socialLayout?: "horizontal" | "vertical";
     appearance?: Parameters<typeof createTheme>[0];
 }
+
+export type ErrorMessage = {
+    title: string;
+    description?: string;
+};
